R0.1

+++ utils/parser.py
@@ -44,7 +44,7 @@ class ConvParser(object):
 				rank = self.ur.readerbench_all(utt1['text'], utt2['text'], lang)['similarityScores']
 				rank['question'] = float(self.ur.is_question(utt1['text']))
 				rank['authorReference'] = self.ur.author_reference(auth1, utt1['text'], auth2, utt2['text'])
-				rank['link'] = float(utt1['genid'] == utt2['ref'])
+				rank['link'] = int(utt1['genid'] == utt2['ref'])
 				
 				ranks.append(rank)
 				print(rank)
